Chore: 유틸 함수 분리 및 알람 메시지에 게시글 제목 추가
